pyproject.toml
[project]
name = "codeadvisor"
version = "0.1.0"
description = "LLM code analyzer and advisor"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.2.1",
    "langgraph>=0.0.62",
    "langchain-community>=0.2.1",
    "langchain-openai>=0.1.8",
    "faiss-cpu>=1.8.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
]
